Move helper files for reppy to assets
