+++ assets/html/report.css
@@ -1,0 +1,408 @@
+/* ── Bijou CSS (vendored) ───────────────────────────────────────────────── */
+@import url(//fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600,700);
+
+a {
+  font-size: 0.95em;
+  text-decoration: none;
+  color: #5f90b0;
+}
+a:hover {
+  color: #406882;
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  font-family: "Source Sans Pro", sans-serif;
+}
+
+.container {
+  width: 1200px;
+  margin: 0 auto;
+}
+
+.pull-left {
+  float: left;
+}
+
+.pull-right {
+  float: right;
+}
+
+.row {
+  min-height: 2em;
+  line-height: 2em;
+  width: 100%;
+}
+
+.span {
+  float: left;
+  display: inline;
+  min-height: 1em;
+  margin: 0 2% 20px 2%;
+}
+.span:first-child {
+  margin-left: 0;
+}
+.span:last-child {
+  margin-right: 0;
+}
+
+.one {
+  width: 8%;
+}
+.two {
+  width: 16%;
+}
+.three {
+  width: 24%;
+}
+.four {
+  width: 32%;
+}
+.five {
+  width: 40%;
+}
+.six {
+  width: 48%;
+}
+.seven {
+  width: 56%;
+}
+.eight {
+  width: 64%;
+}
+.nine {
+  width: 72%;
+}
+.ten {
+  width: 80%;
+}
+.eleven {
+  width: 88%;
+}
+.twelve {
+  width: 100%;
+}
+
+.container:after {
+  content: "\0020";
+  display: block;
+  height: 0;
+  clear: both;
+  visibility: hidden;
+}
+
+.row:before,
+.row:after {
+  content: "\0020";
+  display: block;
+  overflow: hidden;
+  visibility: hidden;
+  width: 0;
+  height: 0;
+}
+
+.row:after {
+  clear: both;
+}
+
+.row {
+  zoom: 1;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-weight: 500;
+}
+
+h1 {
+  font-size: 2.5em;
+}
+h2 {
+  font-size: 1.75em;
+}
+h3 {
+  font-size: 1.5em;
+}
+h4 {
+  font-size: 1.25em;
+}
+h5 {
+  font-size: 1em;
+}
+h6 {
+  font-size: 0.8em;
+}
+
+.table {
+  width: 100%;
+  border-spacing: 0;
+  border-collapse: collapse;
+  text-align: left;
+}
+.table th,
+.table td {
+  border-bottom: 2px #d1d1d1 solid;
+  padding: 8px 12px;
+}
+.table td {
+  border-bottom: 1px #d4d4d4 solid;
+}
+.table.table-striped tbody tr:nth-child(2n + 1) {
+  background-color: #f3f3f3;
+}
+.table.table-bordered {
+  border: 1px #ccc solid;
+}
+.table.table-bordered td,
+.table.table-bordered th {
+  border: 1px #ccc solid;
+}
+
+.button {
+  border: 0;
+  border-radius: 4px;
+  outline: 0;
+  cursor: pointer;
+}
+.button.small {
+  padding: 7px 20px;
+  font-size: 0.95em;
+}
+.button.large {
+  padding: 15px 45px;
+  font-size: 1.1em;
+}
+.button.primary {
+  color: #fff;
+  background: #5f90b0;
+}
+.button.primary:hover {
+  background: #5589ab;
+}
+.button.success {
+  color: #fff;
+  background: #4daf7c;
+}
+.button.success:hover {
+  background: #48a474;
+}
+.button.danger {
+  color: #fff;
+  background: #e6623c;
+}
+.button.danger:hover {
+  background: #e4572e;
+}
+
+.navbar {
+  background: #fff;
+  padding: 10px;
+  margin-bottom: 10px;
+  border-bottom: 1px #ddd solid;
+  height: 30px;
+}
+.navbar h4 {
+  color: #191919;
+  font-weight: 200;
+}
+.navbar li {
+  display: inline-block;
+  padding: 0.3em 15px 0 0;
+}
+.navbar li:last-child {
+  padding-right: 0;
+}
+.navbar.fixed {
+  position: fixed;
+  left: 0;
+  right: 0;
+  top: 0;
+}
+
+.alert {
+  padding: 15px;
+  margin: 10px 0;
+  border-radius: 3px;
+}
+.alert.primary {
+  background: #e8eff3;
+  border: 1px #c5d7e3 solid;
+}
+.alert.success {
+  background: #daefe4;
+  border: 1px #b6dfca solid;
+}
+.alert.danger {
+  background: #fdf4f1;
+  border: 1px #f7cfc4 solid;
+}
+
+@media only screen and (min-width: 1000px) and (max-width: 1199px) {
+  .container {
+    width: 980px;
+  }
+}
+@media only screen and (min-width: 768px) and (max-width: 999px) {
+  .container {
+    width: 768px;
+  }
+}
+@media only screen and (max-width: 767px) {
+  .container {
+    width: 300px;
+  }
+  .container .table {
+    width: 100%;
+    display: block;
+    overflow: auto;
+  }
+}
+@media only screen and (min-width: 480px) and (max-width: 767px) {
+  .container {
+    width: 420px;
+  }
+  .container .table {
+    display: table;
+  }
+}
+
+/* ── GA4GH report custom styles ────────────────────────────────────────── */
+body {
+  margin: 0 auto;
+  font: 16px sans-serif;
+}
+
+h1 {
+  margin-top: 1ex;
+  margin-bottom: 1.5ex;
+}
+
+h2 {
+  margin-top: 1ex;
+  margin-bottom: 1ex;
+}
+
+h3 {
+  margin-top: 1ex;
+  margin-bottom: 1ex;
+}
+
+p {
+  margin-bottom: 1ex;
+}
+
+.centered {
+  width: 100%;
+  text-align: center;
+}
+
+svg text {
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+svg text::selection {
+  background: none;
+}
+
+.axis path,
+.axis line {
+  fill: none;
+  stroke: #000;
+  stroke-width: 2px;
+  shape-rendering: crispEdges;
+}
+
+.axis .label {
+  font: 12px sans-serif;
+  font-weight: bold;
+}
+
+rect.zoom {
+  stroke: steelblue;
+  fill-opacity: 0.5;
+  background-color: #c5d7e3;
+  fill: #c5d7e3;
+}
+
+.dot {
+  stroke: #000;
+}
+
+button.showhide {
+  color: #fff;
+  background: #aaa;
+  margin-top: 15px;
+  padding: 2px 10px;
+  font-size: 0.7em;
+}
+
+button:hover.showhide {
+  background: #999;
+}
+
+.highlighted_metric {
+  font-weight: bold;
+  color: #406882;
+}
+
+table.table {
+  margin-bottom: 0.5ex;
+}
+
+table.table > tbody > tr td.highlighted_metric {
+  background-color: #fafafa;
+}
+
+table.table > tbody > tr td.highlighted_metric_r {
+  background-color: #fffafa;
+}
+
+table.table-striped > tbody > tr:nth-child(2n) td.highlighted_metric {
+  background-color: #fafafa;
+}
+
+table.table-striped > tbody > tr:nth-child(2n + 1) td.highlighted_metric {
+  background-color: #e0e0e0;
+}
+
+.collapsed {
+  display: none;
+}
+
+.hidden {
+  display: none;
+}
+
+input[type="text"] {
+  display: block;
+  margin-top: 10px;
+  width: 100%;
+  font-family: sans-serif;
+  font-size: 18px;
+  box-shadow: none;
+  border-radius: none;
+}
+
+input[type="text"]:focus {
+  outline: none;
+}
+
+input[type="text"] {
+  padding: 10px;
+  border: none;
+  border-bottom: solid 1px #ddd;
+  transition: border 0.2s;
+}
+
+input[type="text"]:focus,
+input[type="text"].focus {
+  border-bottom: solid 1px #aaa;
+}
