+++ assets/html/roc_plot.jinja2.html
@@ -1,9 +1,9 @@
 {%- macro rocplot(roc_name, data) -%}
 <div class="centered"><div id="roc_{{ roc_name }}"></div></div>
 <script>
-    setTimeout(function () {
-        var data = JSON.parse({{ data }});
-        PLOT.roc("#roc_{{ roc_name }}", data);
-    }, 100);
+  setTimeout(function () {
+      var data = JSON.parse({{ data }});
+      PLOT.roc("#roc_{{ roc_name }}", data);
+  }, 100);
 </script>
 {%- endmacro -%}
