--- bin/ga4gh_report_assets/report/__init__.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python
-# coding=utf-8
-#
-# Read GA4GH report files
-#
-# Author:
-#
-# Peter Krusche <pkrusche@illumina.com>
-#
-
-from . import metrics
